提交图片选择以及招聘分类页面

+++ MasterOfHair/MasterOfHair/CustomTool/ImagePicker/Classes/JKImagePickerController.m
@@ -586,8 +586,6 @@ static NSString *kJKAssetsFooterViewIdentifier = @"kJKAssetsFooterViewIdentifier
                                  }];
     
     [picker dismissViewControllerAnimated:NO completion:^{}];
-    
-  
 
     
 }
@@ -596,7 +594,7 @@ static NSString *kJKAssetsFooterViewIdentifier = @"kJKAssetsFooterViewIdentifier
 - (void)startPhotoAssetsViewCell:(JKAssetsViewCell *)assetsCell
 {
     if (self.selectedAssetArray.count>=self.maximumNumberOfSelection) {
-        NSString  *str = [NSString stringWithFormat:@"最多选择%lu张照片",(unsigned long)self.maximumNumberOfSelection];
+        NSString  *str = [NSString stringWithFormat:@"最多选择%lu张照片",self.maximumNumberOfSelection];
         [JKPromptView showWithImageName:@"picker_alert_sigh" message:str];
         return;
     }
@@ -614,7 +612,7 @@ static NSString *kJKAssetsFooterViewIdentifier = @"kJKAssetsFooterViewIdentifier
 - (void)didSelectItemAssetsViewCell:(JKAssetsViewCell *)assetsCell
 {
     if (self.selectedAssetArray.count>=self.maximumNumberOfSelection) {
-        NSString  *str = [NSString stringWithFormat:@"最多选择%lu张照片",(unsigned long)self.maximumNumberOfSelection];
+        NSString  *str = [NSString stringWithFormat:@"最多选择%lu张照片",self.maximumNumberOfSelection];
         [JKPromptView showWithImageName:@"picker_alert_sigh" message:str];
     }
     
@@ -675,7 +673,7 @@ static NSString *kJKAssetsFooterViewIdentifier = @"kJKAssetsFooterViewIdentifier
 - (void)resetFinishFrame
 {
     self.finishButton.hidden = (self.selectedAssetArray.count<=0);
-    self.finishLabel.text = [NSString stringWithFormat:@"下一步(%ld)",(unsigned long)self.selectedAssetArray.count];
+    self.finishLabel.text = [NSString stringWithFormat:@"下一步(%ld)",self.selectedAssetArray.count];
     [self.finishLabel sizeToFit];
     
     self.finishButton.width = _finishLabel.width+10;
